fix(rar5): wrap window positions instead of masking them (M4.3)
RAR7 dictionaries are not powers of two. The compression-info word carries a
5-bit fraction on top of the 5 dictionary bits (`winSize += winSize/32*frac`,
`d861246:arcread.cpp:867-882`), which junrar has parsed since M4.1, so a
declared window such as 0x42000 is already reachable — and reachable well below
the 1 GB engine capability, not only above 4 GB.

Every window position was folded with `& maxWinMask` where `maxWinMask` is
`size - 1`. On a non-power-of-two window that mask wraps at the size rounded
down to a power of two, addressing window space that was never written: a real
`-md=264k` archive (verified `All OK` by unrar 7.23) failed extraction with
`ArrayIndexOutOfBoundsException: Index 263878 out of bounds for length 262144`.

Port `WrapUp`/`WrapDown` from `d861246:unpack.hpp:416-435`, whose own comment
says they exist to replace `&MaxWinMask` for non-power-of-two windows, and take
window positions to `long` (C++ `size_t`, `unpack.hpp:277-283`) so the
arithmetic has room for the 64 GB windows M4.3 goes on to allow. `AddFilter`
keeps upstream's `% MaxWinSize` rather than a wrap, because a malformed block
start can be many times the window size (`unpack50.cpp:228-233`), and
`Unpack5Filter.blockStart` widens with it (`unpack.hpp:150`). The engine's own
`Unpack5Window` API becomes long-indexed and gains the `run`/`bufferAt`/
`offsetAt`/`copyOut` seam the write-out and filter-copy paths need to stop
assuming the window is one contiguous array.

For a power-of-two window every site is arithmetically identical to the mask it
replaces, so no existing behaviour moves.

Tests: the fractional-window extraction is RED first (frozen byte-identical at
`git hash-object 75e131c96b3270abca611cefc57b9d64530c3edb`), plus a sweep over
five fraction values that move the wrap point across the payload. Mutation
probes exposed three genuine gaps in wrap coverage that predate this chunk and
are closed here: a match straddling the window end from either side, a filter
block straddling it, and a filter start folded from beyond the window.

Refs #35

+++ src/main/java/com/github/junrar/unpack/Unpack5Filter.java
@@ -10,15 +10,16 @@ import com.github.junrar.unpack.decode.Compress;
  * One RAR5 filter queue entry (M3.8, issue #29; unrar {@code 8f437ab:unpack.hpp} struct
  * {@code UnpackFilter}). Mirrors the RAR3 {@link UnpackFilter} shape minus the VM fields —
  * RAR5 dropped the VM entirely, filters are just a type tag plus a window range (manual
- * &sect;5.4). {@code blockStart} and {@code blockLength} are C++ {@code uint}: raw header
- * values until {@code AddFilter} masks the start into the window, so callers comparing them
- * use unsigned semantics (no-go C15).
+ * &sect;5.4). {@code blockLength} is a C++ {@code uint} and stays {@code int}; {@code blockStart}
+ * is zero-extended to {@code long} on read, because {@code AddFilter} folds it into a window
+ * position ({@code size_t}, up to 64 GB) — {@code d861246:unpack.hpp:150}, where the field was
+ * widened from {@code uint} to {@code size_t} for exactly that reason (no-go C15).
  */
 class Unpack5Filter {
 
     private int type = Compress.FILTER_NONE;
 
-    private int blockStart;
+    private long blockStart;
 
     private int blockLength;
 
@@ -34,11 +35,11 @@ class Unpack5Filter {
         this.type = type;
     }
 
-    public int getBlockStart() {
+    public long getBlockStart() {
         return blockStart;
     }
 
-    public void setBlockStart(final int blockStart) {
+    public void setBlockStart(final long blockStart) {
         this.blockStart = blockStart;
     }
